Paranneltu tekstikappaleiden tuontia

+++ eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/service/impl/PerusteenOsaViiteServiceImpl.java
@@ -128,7 +128,8 @@ public class PerusteenOsaViiteServiceImpl implements PerusteenOsaViiteService {
             throw new BusinessRuleViolationException("Sisällöllä on lapsia, ei voida poistaa");
         }
 
-        if (viite.getPerusteenOsa() != null && viite.getPerusteenOsa().getTila().equals(PerusteTila.LUONNOS)) {
+        if (viite.getPerusteenOsa() != null && viite.getPerusteenOsa().getTila().equals(PerusteTila.LUONNOS)
+                && findViitteet(perusteId, id).size() == 1) {
             PerusteenOsa perusteenOsa = viite.getPerusteenOsa();
             perusteenOsaService.delete(perusteenOsa.getId());
         }
@@ -193,6 +194,12 @@ public class PerusteenOsaViiteServiceImpl implements PerusteenOsaViiteService {
 
     @Autowired
     private PerusteRepository perusteet;
+
+    private List<PerusteenOsaViite> findViitteet(Long perusteId, Long viiteId) {
+        PerusteenOsaViite viite = findViite(perusteId, viiteId);
+        List<PerusteenOsaViite> viitteet = repository.findAllByPerusteenOsa(viite.getPerusteenOsa());
+        return viitteet;
+    }
 
     private PerusteenOsaViite findViite(Long perusteId, Long viiteId) {
         Peruste peruste = perusteet.findOne(perusteId);
